+++ src/extension/utils/fileUtils.ts
@@ -2,12 +2,16 @@
 import { existsSync } from 'fs';
 import path = require('path');
 export function getUnSaved(scriptFileName: string) {
-    const dirname = path.dirname(scriptFileName);
-    const extname = path.extname(scriptFileName);
-    const basename = path.basename(scriptFileName);
+    if (!existsSync(scriptFileName)) {
+        return scriptFileName;
+    }
+    const { dir, base: baseName } = path.parse(scriptFileName);
+    const indexStart = baseName.indexOf('.');
+    const fileNameWithOutExt = baseName.substr(0, indexStart);
+    const ext = baseName.substr(indexStart);
     var i = 0;
-    while (existsSync(path.join(dirname, `${basename} [${i}].${extname}`))) {
+    while (existsSync(path.join(dir, `${fileNameWithOutExt} [${i}]${ext}`))) {
         i++;
     }
-    return path.join(dirname, `${basename} [${i}].${extname}`);
+    return path.join(dir, `${fileNameWithOutExt} [${i}]${ext}`);
 }