+++ src/renderer/renderer.tsx
@@ -2,16 +2,18 @@
 import { FunctionComponent, h } from 'preact';
 import { StateUpdater, useEffect, useState } from 'preact/hooks';
 import { v4 as uuidv4 } from 'uuid';
-import { DothttpExecuteResponse } from '../common/response';
+import { RendererContext } from 'vscode-notebook-renderer';
+import { DothttpExecuteResponse, MessageType } from '../common/response';
 var stringify = require('json-stringify-safe');
 // import * as Search from 'vscode-codicons/src/icons/search.svg';
 
-export const Response: FunctionComponent<{ response: Readonly<DothttpExecuteResponse> }> = ({ response }) => {
+export const Response: FunctionComponent<{ response: Readonly<DothttpExecuteResponse>, context: RendererContext<any> }> = ({ response, context }) => {
     const [activeIndex, setActive] = useState(0);
     const [searchKeyword, setSearchKeyword] = useState('');
     const uuid = uuidv4();
     const searchBarId = `search-bar-${uuid}`;
     const searchButtonId = `search-button-${uuid}`;
+    const saveButtonId = `save-button-${uuid}`;
 
     let darkMode = document.body.getAttribute('data-vscode-theme-kind')?.includes('dark') ?? false;
 
@@ -58,6 +60,8 @@ export const Response: FunctionComponent<{ response: Readonly<DothttpExecuteResp
                 scriptResultExists={scriptResultExists}
                 generatedProperties={propertiesGenerated} />
             <span class='tab-bar-tools'>
+                <button id={saveButtonId} class='search-button' title='Save response' onClick={() => saveResponse(context, response)}>Save Response</button>
+                <button id={'generate-lang-${uuid}'} class='search-button' title='Generate Language' onClick={() => generateLanguage(context, response)}>Generate Language</button>
                 <input id={searchBarId} placeholder='Search for keyword'></input>
                 <button id={searchButtonId} class='search-button' title='Search for keyword' onClick={() => handleSearchForKeywordClick(setSearchKeyword, searchBarId)}>
                     {/* <Icon name={Search} /> */}
@@ -222,3 +226,12 @@ const searchForTermInText = (text: string, searchKeyword: string) => {
         })}
     </span>;
 };
+
+function saveResponse(context: RendererContext<any>, response: any): void {
+    // console.log(response);
+    context.postMessage!({ "response": response, "request": MessageType.save, })
+}
+function generateLanguage(context: RendererContext<any>, response: Readonly<DothttpExecuteResponse>): void {
+    context.postMessage!({ "response": response, "request": MessageType.generate, })
+}
+